requirements.txt
audio-recorder-streamlit==0.0.10
ipykernel>=6.29.5
langchain>=0.3.17
langchain-ollama>=0.2.3
langgraph>=0.2.68
loguru
ollama>=0.4.7
openai-whisper
pyprojroot>=0.3.0
python-box>=6.0.2
PyYAML>=6.0.2
streamlit==1.42.0
streamlit-float==0.3.5
torch>=2.6.0
torchaudio>=2.6.0
transformers>=4.48.3
-e .